requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
pytz==2024.1
python-dotenv==1.0.1